use Slider in ScrolledView and fix/optimize Slider
This is the start at Slider inside a Scrolled View instead of having Scrolled
view draw its own idea of a slider. This is clunky now, but more options on
Slider will give the intended look.

This also cleans up and fixes many minor bugs with Slider, hopefully more than
it creates.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>