experimental: make widgets copyable and moveable
This is a very experimental at making widgets copyable and moveable, but being
able to do this is probably to complex to make proper use of due to exceptions.

For example, this allows copying widgets into a vector, but those widgets will
no longer be attached to any previous parent they have.  However, if they are
moved into the vector, they will of course still maintain all of their existing
properties.  Having a user make this distinction is nasty.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>