standardize widget object lifetime
This moves to two simple APIs for management: user managed or EGT managed with
shared_ptr.

Several bugs in grid are resolved (deleting children) as well as an API change
to remove the align parameter when adding widgets.  Instead, the new align
utility functions should be used, or explicitly Widget::set_align() as always.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>