fix vertical alignment of text
Text is drawn in two different ways vertically - font aligned and text
aligned. Buttons and that want to center the actual text displayed, use text
aligned. TextBox, which wants to align based on the text baseline, does it by
font. This change accounts for the difference of the two, even if not obvious.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>