fix pkg-config private/public libs
This fixes several issues with already linked vs need-to-link libs exposed
through pkg-config.

There general rule is if a dependent library is exposed in the API then it needs
to go in the Requires section.  If it is not, then we need to put it in
Requires.private and directly link against it.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>