inputlibinput: remove screen hard coded size
The size of the screen is needed to compute the x,y coordinates.
libinput_event_touch_get_x() should return an absolute coordinate in mm.
Depending on how the device reports the coordinates, it can also be an
absolute raw value. For these reasons, prefer the transformed which
returns screen coordinates.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>