Sizer: super_rect() rework
Change behavior of horizontal and vertical sizers.

The expected behavior for an horizontal sizer is that its height is the maximum
height among its children. Its width is the addition of the width of its
children excepted the ones which are expanded horizontally as their size depends
on the free space available in the sizer. If the sizer is horizontally
expanded, don't reduce its width in order to not remove the extra free
space which will be used by expanded children.

It's the very same behavior for vertical sizers excepted that height and
width logic is inverted.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
[style code]
Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>