hardware overlay z-order is not chronological order by default
The order changes by device.  Typically, the z-order of hardware overlay planes
is:

    OVR2 > OVR1 > HEO > BASE

There are limited ways in hardware to change this, but this is the default
provided by the driver. This change simply adds a hard coded array to respect
this order, but it means that the first overlay allocated without a preference
is going to be the HEO overlay plane. This change also consolidates the code to
allocate the plane into a single loop.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>