gstappsinkimpl: fix the video freeze on pause.
All of the pipeline elements were changing from PLAYING to PAUSED and
PAUSED to PLAYING without any pending state. however incase of fakesink
the state remains at PLAYING and a pending PAUSED state which is because
it attempts to change it asynchronously. setting async & sync property
to false fix this issue.

Signed-off-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>