rework and document Gauge classes
This documents and further completes the API for the Gauge classes.  In addition,
the examples using these classes are updated. The big change here is converting
the Gauge classes to proper Widget/Frame based classes.  Also, gauge layers can
either be individual objects, or cover the entire gauge.  Handling for
positioning and sizing of SVG based gauge layers is further improved with the
use of the SvgImage class.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>