support --enable-simd in make dist/distcheck
This is overly complicated and seems to be a pretty frail solution.  Create the
build directory and then run cmake referencing where the srcdir is. Simd then
wants to make changes in the srcdir and make distcheck doesn't allow it.  So,
some errors are printed, but all seems to work.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>