examples/easing: don't perform parent_layout() when changing label text
This causes a nasty problem that causes studders in the otherwise smooth
animation.  Changing the CPU label text causes a parent_layout() of the whole
window which is expensive as a gap in animation.  The widget flag no_layout
prevents this, but now we need to manually deal with the label size as
construction.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>