StaticGrid: force widget position into the cell
The alignment algorithm won't managed the position of the widget in
accordance with the bounding if there are no alignment flags.
Let the StaticGrid manage the position of the widget according to the
bounding. Then if the position has be be changed in regards to the
alignement constraints, it will be managed by the alignment algorithm.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>