ComboBox: fix bug preventing to close the popup
The popup of the combox can't be close if the item selected is not the
first one. In fact, it's closed and redraw.
The root cause is the hide() call on on_selected_changed event. When
a click happens on the ComboBox, it causes a show_modal() call. During
this call, the popup is set a the modal window and the ComboBoxPopup
show() method is called. It reconstructs the list and selects the item
selected previously then show the Popup. Unfortunately if the item
selected is not the first one, it causes an on_selected_changed event
that hides the ComboBoxPopup before calling the Popup show() method.
So this popup is showed but it has lost its modal window attribute. The
consequence is that click outside the popup are no longer redirected to
the ComboBox.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>