ListBox: add new on_selected signal
The first item of the list is selected by default. Thus, when there is a
click on it, nothing happens as the item selected has not changed. Now a
separate event can be used to see any time an item is selected, which is
different than on_selected_changed signal.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>