Font: support loading fonts directly without Fontconfig
This allows a URI to be used to specify a font face and FreeType is directly
used instead of going through Fontconfig.  This is a significant performance
increase and also paves the way to loading fonts from other locations.

This also adds a configure option to explicitly disable the use of Fontconfig.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>