NeedleLayer: add an optimization to limit the damage() rectangle when redrawing
This calculates a tighter dynamic rectangle around the rotated needle to limit
the amount of surface area drawn.  This is done by rotating the needle image
rectangle to the specified angle, as is done when drawing, and then calculating
the super rectangle that is axis aligned.  This rectangle is then used to
damage the NeedleLayer instead of damaging the whole layer.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>