configure.ac: rework handling of setting CFLAGS/CXXFLAGS
No longer explcicity set -g or -O flags. This is not something we should be
maintaining and it conflicts with a user trying to specifiy it. Allow users to
configure with their own flags generally instead of forcing flags.  Properly
check and add compiler flags.  Better handling of CFLAGS in addition to
CXXFLAGS.  Use a better method to set and check the c++ standard (adds to CXX
instead of CXXFLAGS).  Show CFLAGS and LDFLAGS in summary in addition to CXXFLAGS.

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>