Grid: rework of the layout
The border of the widget was also used as the grid border i.e between
the cells.
So if someone doesn't want to display the external border, there is no
way to add empty space between the cells.

This patch changes the layout of grids. Margin, border and padding have
the same behavior as other widgets. If there is a need to put extra
space between the cells, use the horizontal or vertical space. This
change makes the border parameter from the grid constructors useless.
Same for the grid_flags which were only used to show the inner border or
not. The notion of inner or grid border is away.

SelectableGrid changes as well, now you can specify the size of the
highlight border with selection_highlight(). This border is drawn within
the grid cell. It means it will be drawn on top of the cell widget. If
you don't want to draw on top of the widget, then set a margin for the
widget.

Serializatio/deserialization have been updated according to these
changes.

Finally to not introduce compilation or behavior breakages, widgets or
examples that used the grid border as a margin and grid space have been
fixed.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>