slider: removal of the origin_opposite flag
As sliders use start and end instead of min and max and it's allowed to
get end > start, the origin_opposite flag is no longer needed. This flag
was used by the ScrolledView, when scrolling from the bottom to the top
with the finger, the handle of the slider goes in the opposite
direction.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>