CameraImpl: use the user requested box if available
If the user specified a box for the camera window, use this box instead
of the current box. In the past, the window size was not updated when
scaling happens. It was the cause of input events not caught when
upscaling happened. If we want to use the hardware to scale the camera
video, we have to use the user requested box.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>