bump libtool version-info to prepare 1.2 release
bump libtool version-info from 2:0:0 to 3:0:0. API changes are addition
for this release or removal of logic constraints about parameters for
RangeValueWidget. It could be considered to set version info to 2:1:0
but we may not support this use case:
- Programs using the previous version may use the new version as drop-in
replacement, and programs using the new version can also work with the
previous one. In other words, no recompiling nor relinking is needed.
In this case, bump revision only, don’t touch current nor age.

A program using the new version may not work with the previous one
because of the removal of logic constraints about parameters for
RangeValueWidget, start has no longer to be greater than end.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>