git_version: fix particular build use case
Fix a build use case discovered with Buildroot. Buildroot creates a
source archive from the git repo. If this archive is extracted in a git
repo, the commit id of this git repo will be used as the git version of
EGT. So if the top-level directory is not the current directory, discard
the commit id.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>