AnimationSequence: improve the API
A sub animation can't be added several times to a sequence. It can be
annoying especially when users want to add the same animation delay
several time. This constraint is now removed.
The side effect is that the behavior of the remove method is no longer
clear. Should it remove all the occurences of the animation or only one?
Update the documentation to clarify this point, it will remove the first
occurence.
To be more convenient add removeAt and removeLast methods.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>