Widget: fix margin, border, padding increase
Margin, border and padding should reduce the content area when they are
increased. This is not the case if autoresize is set and the current size
of the widget is the minimal size hint, the widget box is increased as
the content area can't be shrinked. So discard the increase in this
case.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>