ImageButton/Label: change min size hint behavior
The expected behavior for ImageButton/Label, if no size is specified, is
to have a widget matching the size of the image, not a bigger one.
The default size is based on the screen size. This default size is used
as the minimal size. In some cases, the default size is bigger than the
size of the image to display, in this case, shrink the size to the
image.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>