Widget: devirtualize Widget::deserialize()
The deserialize() private method is only called from constructors that
take Serializer::Properties& props as parameter.

However the deserialize() method is never called with a pointer or a
reference to a base class so the overide of the derived class is
actually executed. Therefore, the virtual mechanism is never used.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>