Widget: add deserialize_leaf()
The deserialize_leaf() method should be called at the end of deserializer
constructors T::T(Serializer::Properties& props, bool is_derived), if
and only if its 'is_derived' parameter is false.

Indeed, the additional 'is_derived' boolean parameter tells the constructor
whether it is executed as the constructor of a base class (is_derived ==
true) from another constructor of a derived class.

As a matter of fact, the deserialize_leaf() method is supposed to be
called once the object is completely constructed, so calls of virtual /
overridden methods actually execute the relevant and expected code.

This allows for instance to execute the proper version of the show()
method when deserializing a widget.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>