SvgImage: copy the URI when creating an Image from a SvgImage
Add a new private Image constructor, for internal use only, that
calls the Image(shared_cairo_surface_t) constructor and copies the
additional 'const std::string& uri' parameter into the Image::m_uri
member.

The SvgImage class becomes a friend of the Image class so it can make
use of this new private Image constructor.

The purpose of this patch is to keep trace of the URI when an egt::Image
is created from an egt::SvgImage so the MGC can properly save the 'uri'
property in the JSON screen file.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>