ImageHolder: draw the widget text if m_show_label is true
Test show_label() rather than widget.text().empty() to decide whether a
text should be drawn aside the image.

Currently, the EGT supports image auto-scaling in ImageHolder only when no
text is drawn. Then, in the MGC, it is more intuitive for users to
uncheck the 'Show Label' box rather than clearing the widget text in the
object editor to allow the 'Auto Scale' property to be enabled.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>