docs: warn that the default VirtualKeyboard can be instanciated only once
Due to the way pre-defined PanelKeys are designed (static variables),
they can be instantiated only once. The second instance will cause an
assertion when it will be added to a parent as there is already a parent
due to the first instance. A consequence is that only one default
VirtualKeyboard can be instantiated.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>