Slider: add the 'inverted' flag
Swapping 'start' and 'end' of the value range has an impact only on the
slider value and labels whereas setting the 'inverted' flag also changes
the way the slider is drawn.

Therefore, defaut sliders progress from the left to the right for
horizontal sliders and from the bottom to the top for vertical sliders
while inverted sliders progress from the right to the left for
horizontal sliders and from the top to the botton for vertical sliders.

Also fix ScrolledView where the horizontal slider had been inverted by
mistake by the commit which had removed the 'origin_opposite' flag.

Fixes: 1ce09d683c99 ("slider: removal of the origin_opposite flag")
Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>