ScrolledView: allow offsets to be modified even if no scroll bar is visible
Even without testing whether any scroll bar is visible,
the ScrolledView::offset() method checks that the new offsets belong to
the offset range anyway.

Therefore, we can safely remove the useless *scrollable() tests and so,
we can let the user shift the drawn area of the ScrolledView with the
mouse (click then drag).

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>