Serializer: remplace vectors with lists for properties and attributes
The main advantage of std::vector on std::list is the direct access to a
random element but this feature is not used.

Indeed, the serialization and deserialization of widgets only iterate on
properties and attributes containers and erase elements.

Then, when iterating, both std::vector and std::list have a linear
complexity O(n). However, erasing an element from an iterator is done
in a constant time O(1) for std::list whereas it is done in a linear time
O(n) for std::vector.

Properties are erased one by one from the Widget::deserialize() virtual
method.

Hence, std::list is better suited than std::vector to store
Serializer::Properties and Serializer::Attributes.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>