ImageButton/Label: change the way show_label is managed
show_label from ImageHolder is by default set to false. If an
ImageButton/Label is created without a string, it is set to false. This
behavior is annoying in the case of the composer which uses the default
constructor and after set the text.
Now, show_label is set to true when text is set with a non-empty string
and false otherwise. User can still change the show_label status if the
want after having set the text.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>