TextBox: fix textbox text x & y positions
The m_rects x & y values are calculated only on first draw due to which
textbox text is not getting changed when textbox is moved/resized to
different position/size and at present prepare_text method update texts
only if layout changes. so as a fix calling prepare_text method to update
text if layout or box properties changes.

Signed-off-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Reviewed-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>