Sound: fix bug when using ifstream
When ifstream is used to handle the sound file, the sound can be played
only once. Once played the m_impl->in condition becomes false, that
prevents to play the sound several time as it's the first condition
checked in the play method.
The ifstream bool operator returns false if failbit or badbit are set.
Usually, the failbit is set after reading the file once the end of
file is reached. Usually further operations on the stream may be
possible. If badbit is set, it may involve the loss of integrity of the
stream. So clear the iostate only if badbit is not set.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Reviewed-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>