fix git version generation
Git no longer allows, by default, to run git commands if we are not the
owner of the git tree. On side effect is that during a sudo make install
command, as we are root, there is a mismatch of ownership for the git
tree. Git commands return an error and the git_version script returns
n/a.
To fix this behavior, if we detect that make is run as root and the git
tree is owned by another user, switch to this user to run the
git_version script.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>