Image: fix alignment when keep_image_ratio is set
The alignment algorithm doesn't force the orig rectangle inside the
bounding box if there is no alignment flag. As the alignment flag is
removed to preserve the new size computed to preserve the ratio of the
image, we should face this case. So it's important to not rely on the
bounding rectangle and to set the position of the orig rectangle.

Fixes: b334fe0cc451 ("Image: fix align() to force rescaling when keep_image_ratio is enabled")
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>