configure.ac: fix generating of libegt.dll on windows machine
while building libegt on windows10 machine using MSYS2, observed
warning message saying SDL2main.dll not found. Due to which only
libegt.a file is getting generated and not generating libegt.dll.
so removing -lSDL2main from SDL2 LDFLAGS fix this issues. The
SDL2main provides SDL_main(), which is a global entry point for
all SDL apps. You are not required/forced to use it, but is presence
is based on the variety of the systems SDL supports. Windows uses
WinMain() and Linux uses main(). so SDL_main() gets replaced by main()
when SDL2main linked and libegt does not have a main function.

Signed-off-by: Sandeep Sheriker M <sandeep.sheriker@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>