Revert "ImageButton/Label: change the way show_label is managed"
This reverts commits 2a37aeff1713a4cf8f8b523cb949e8e363a21b75 and
51b46cbfb42d935a8845abf354b6c98250d6c7fa.

These patches created an issue in the MGC when loading a project with
ImageLabel or ImageButton widgets.

Indeed, when a project is loaded, for each widget, the MGC first creates
a default widget then sets all its properties one by one in alphabetical
order.

Therefore, for ImageLabelWidgets and ImageButtonWidgets saved with
'imageholder_show_label' set to 'false' and with 'text' set to a
non-empty string, when reloaded, the 'imageholder_show_label' property
was set first to 'false', then 'text' was set next, calling
Image*::text(), which in turn called Image*::show_label(true) because of
the non-empty text string. At the end, 'imageholder_show_label' was set to
'true' when it should have been set to 'false'.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>