XmlWidgetSerializer: avoid a useless copy of std::vector<>
Pattern::StepArray is a std::vector<>. Also, Pattern::steps() returns a
'const StepArray&', hence a reference.

So, replace 'auto' with 'const auto&' to avoid copying this std::vector<>
but instead keeping a const reference on it.

Again, replace 'auto&' with 'const auto&' when iterating on 'steps' as we
don't modify any element of the 'steps' vector.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>