OstreamWidgetSerializer: fix the serialization of the TopWindow
Since the serialization of the global theme, palette and font has been
introduced, the serialization of the TopWindow is broken.

Indeed, the children of the TopWindow appears at the same level as the
TopWindow instead of one level deeper: more precisely the TopWindow and
its children all appears as sibling children of the new "egt" root node.

1 - One too many StringNode* in 'm_impl->stack'

The first issue comes from the number of StringNode* in 'm_impl->stack',
which has changed after commit f08f51e20ac6.

Before commit f08f51e20ac6 ("serialize: serialize global properties"),
when the OstreamWidgetSerializer::add() method was called first, hence
with the 'widget' parameter pointing to the TopWindow and 'level' set to
0, the 'm_impl->stack' stack still contained only one StringNode*,
'{&m_impl->doc}', before 'widget->serialize()' was called.
Therefore, when the TopWindow was being serialized, the add() method was
called again on each of its children with 'level' set to 1.
So when entering add() with 'widget' set to any TopWindow child and
'level' set to 1, 'level' was compared with 'm_impl->stack.size() - 1',
ie 0.
Consequently 'advance' was 'true' so 'm_impl->current', which pointed to
the TopWindow, was pushed back to 'm_impl->stack'.
Later, as the TopWindow child node was appended to the children of
'm_impl->stack.back()', which pointed to the TopWindow node, this child
node was inserted as a child of the right node. At the end, the TopWindow
and children were all serialized correctly.

However, after commit f08f51e20ac6, when to add() method was called first,
still with the 'widget' parameter pointing to the TopWindow and 'level'
set to 0, the 'm_impl->stack' now contained two StringNode*,
'{&m_impl->doc, egt}', just before 'widget->serialize()' was called.
Therefore, when the TopWindow was being serialized, the add() method was
called again on each of its children with 'level' set to 1.
So when entering add() with 'widget' set to any TopWindow child and
'level' set to 1, 'level' was compared with 'm_impl->stack.size() - 1',
which this time was equal to 1 instead of 0 previously.
Consequently 'advance' was no longer 'true', so 'm_impl->current' which
pointed to the TopWindow, was not pushed back into 'm_impl->stack'.
Later, as the TopWindow child node was appened to the children of
'm_impl->stack.back()', which pointer to the 'egt' node, this child was
incorrectly add as child of the 'egt' node instead of child of the
TopWindow node.

2 - Broken add_node() / previous_node() methods

In order to work properly, the OstreamWidgetSerializer::and_node() method
was implemented assuming the 'm_impl->stack' stack already contained two
items when called with m_level set to 0: '{&m_impl->doc, egt}'.

However, we have seen just above in part 1, that 2-element stack had
broken the serialization of the TopWindow.

Therefore, we now avoid calling those broken add_node() and
previous_node() methods and instead directly create nodes for the global
palette and font, without modifying neither 'm_level' not 'm_impl->stack'.

3 - Print 'm_impl->doc' instead of 'm_impl->doc.children.front()'

Before commit f08f51e20ac6, the StringNode* hierarchy was:

doc
 ┕ TopWindow

Only the TopWindow and its descendants were printed, ignoring doc,
therefore starting from 'm_impl->doc.children.front()'

Now, the StringNode* hierarchy is:

doc (nammed 'egt')
 ┝ global theme
 ┝ global palette
 ┝ global font
 ┕ TopWindow

We now want to print from the 'egt' node, therefore from 'm_impl->doc'.

Fixes: f08f51e20ac6 ("serialize: serialize global properties")
Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>