Serializer: add new begin_child() and end_child() virtual methods
The new begin_child() and end_child() methods extend the API of the
Serializer class to provide a generic mean to add children nodes when
serializing objects such as widgets.

The {Ostream,Xml}WidgetSerializer::add() methods are reworked to be
implemented with Serializer::begin_child() / Serializer::end_child(),
hence better sharing the source code.

Also, in both {Ostream,Xml}WidgetSerializer implementations, we no longer
use the 'm_impl->stack' stack and 'm_level' integer: instead we create
local variables (taking advantage of the call stack), of the new
'Serializer::Context*' type, to save and restore the serializer context in
order to manage the node hierarchy properly.

This simplifies the source code and makes it easier to understand and to
use.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>