StaticGrid: fix the serialization/deserialization of children
StaticGrid with blank cells in the middle of the grid were not
serialized correctly. Indeed, the cells coordinates were not serialized,
hence could not be restored at the right location when deserialized.

Actually, all children were serialized the same way in
Frame::serialize(), which is of course not aware of the StaticGrid
cells.

So, create a new serialize_children() virtual method that can be
overridden, for instance in StaticGrid, to tune the way children are
serialized.

Accordingly, adpat the way children are deserialized, introducing a new
deserialize_children() virtual method, that should also be overridden
when needed to match what is done in serialize_children().

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>