GstAppSinkImpl: enlarge caps for the GstURIDecodeBin element
The video player example is broken for some time on PCs. It works only
if the inner decoder of GstDecodeBin has a pad with caps video/x-raw and
no caps features.

The GStreamer pipeline of the video player uses the GstURIDecodeBin
element. It contains the GstDecodeBin element which instantiates
dynamically a decoder.
Until now, the configuration of the machines causes instantiating the
mpeg2dec decoder. Depending on the packages installed on the machine,
other decoders can be used, such as avdec_mpeg2video and nvmpeg2videodec.
With those decoders the video player doesn't work.

The reason of this bug are the caps set for the GstURIDecodeBin. They're
propagated to GstDecodeBin and are set to video/x-raw. Contrary to the
pipeline where, usually, intersection of capabilities are looked
for to find compatible elements to link, GstDecodeBin output must
include the capabilities of the decoder.
The mpeg2dec caps are video/x-raw so they are included in video/x-raw.
The avdec_mpeg2video and nvmpeg2videodec caps have features. For
instance the caps of avdec_mpeg2video are
video/x-raw;video/x-raw(format:Interleaved) so they are not included in
video/x-raw. To link the decoder sink pad to the GstDecodeBin src pad,
the GstDecodeBin caps must be set to video/x-raw(ANY), otherwise the
decoder caps are not a subset of the GstDecodeBin caps.

Even if there is no issue with audio right now, add ANY as well to be
safe.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>