Windows support: move to the C++17 std::filesystem API
The source code of exe_pwd() was suited for Linux but not for Windows:
"/proc/self/exe" only exists on linux-like systems but not on Windows.

In order to make it work, a work-around was added in the Windows version
of the readlink() function: if its 'path' parameter was "/proc/self/exe",
then a special code was executed; this was an ugly trick this patch gets
rid of.

Besides, that special code itself was broken and likely never tested at
all. Indeed, the documentation of GetModuleFileNameA() states that:

"If the function succeeds, the return value is the length of the string
that is copied to the buffer, in characters, not including the terminating
null character."

Therefore, it doesn't make any sense to convert this length with
std::to_string() then to return this string conversion of an integer as
the resolved file path.

Moreover, there was another mistake in the Windows implementation of
readlink(): sizeof() was used instead of strlen() to compute the length
of the string returned by _fullpath(). As a matter of fact, the
_fullpath() function returns a 'char *', therefore
'sizeof(_full_path(...))' was equivalent to 'sizeof(char *)', which was
not the same as 'strlen(_full_path(...))'.

So, in order to fix all those issues on the Windows implementation and
to have a more portable and easier to maintain source code, move to the
C++17 std::filesystem API.

That is to say, readlink(), abspath() and cwd() are now implemented with
the read_symlink(), absolute() and current_path() functions of the
std::filesystem namespace introduced with C++17.

Besides, GetModuleFileNameA() is still used on Windows to get the path
of the executable linked with libegt but now it is called directly from
exe_pwd().

Finally, introduce the dos2unix_path() function to replace back-slashes
'\\' with slashes '/' in paths on Windows.

This way, paths are uniformized, using only slashes '/'. Indeed, even for
Windows, Qt uses slashes '/' as directory separators.

For instance:
"C:/msys64/home/user/Sources/egt/icons".

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>