CameraImpl: keep an internal list of devices
When a monitor is started, there are notifications for the devices
discovered. With the camera example, it causes to get the dialog box
prompting there is a new device connected. There is no way to flush
these notifications.
So when the CameraImpl is constructed, list the devices connected and if
a notification for a device already connected is received, discarded it,
don't generate an event.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>