Painter: add multiline support to the text_size method
The text size method doesn't take into a account line returns in
strings, it assumes there is a single line. So both width and height
computed are incorrect.
As Cairo doesn't provide helpers for that, the size has to be computed
manually finding the longest line to get the width and the number of
lines to get the height.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>