move some Frame features to Widget
This patch is the first of a serie that aims to rework Frame
inheritance.

Now, widgets that are composed of several widgets inherits from Frame to
benefit from child handling. For instance, the Scrollwheel is composed
of a StaticGrid for the layout plus two Buttons and a Label. Instead of
being a StaticGrid, the Scrollwheel must have a StaticGrid. Being a
StaticGrid makes it easily corruptible as it inherits the public add and
remove methods from Frame.

Let's try to break down this move in several patches instead of a big
one. This first patch moves draw, handle and layout
methods from Frame to Widget and all the necessary code to make it
compile.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>