TextBox: delete the whole selected text when DEL or BACKSPACE is pressed
Even when the selected text contained more than one character, only a
single one was deleted; if DELETE was pressed then the character
following the cursor, if any, was deleted, whereas if BACKSPACE was
pressed then the character preceding the cursor, if any, was deleted.

Change this behavior so that if a selected text exists, then this text
is fully erased instead of a single character.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>