TextBox: fix a damage issue when deleting characters
To reproduce the issue fixed by this patch, create a TextBox, clear its
text and click on it so it has the focus and its cursor blinks.

Then press 'q' twice so the TextBox text is "qq". Wait for the cursor to
be drawn then press BACKSPACE: The 2nd 'q' is not totally erased; its
right side wasn't part of the damage area, hence leaving a vertical bar
of 1-pixel width.

This is due to a rounding issue in split() when computing:
tail_rect.width(tail_rect.width() - head_te.x_advance); [1]

For instance, tail_rect.width() could be at first 63 and head_te.x_advance
was 54.00000000000001; then [1] resulted in tail_rect.width() being
updated to 8 instead of 9.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>