TextBox: fix the text token widths due to the layout boundaries
When the rectangle of a text token crossed the right side of the
'boundaries' rectangle during the layout computation but was too large
anyway to fit into the next line, its width was wrongly truncated based
on the width of the 'boundaries' rectangle: its x coordinate was still
correct though, crossing the boundary width if necessary, whereas its
width was truncated.

So once the layout is computed, update only the x and y coordinates of
TextRect rectangles but don't change their width or height.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>