TextBox: stop adding new tokens after '\n' if 'multiline' flag is not set
In tokenize(), if the 'multiline' flag is not set, stop appending new
token when the first '\n' is found, since those tokens wouldn't be drawn
anyway.

Hence, in draw_text(), we no longer need  to stop drawing tokens when the
first '\n' is reached and 'multiline' is not set.

Also, the TextBox::TextFlags parameter of draw_text() is no longer used,
therefore it is removed.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>