TextBox: don't damage the text or the cursor if not visible
The text could be damaged even if not visible, ie if intersecting the
content area. Change that so only the visible text can create damages.

Do the same with the cursor, except that the widget box should be used
instead of its content area when intersecting with the cursor rectangle.
Indeed, the cursor may cross the content area boundaries, especially when
positioned at the very beginning of a line.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>