TextBox: fix cursor position after a line break which is not \n
When the layout forced a line break after a token which was not a \n,
get_cursor_rect() wrongly positioned the cursor at the end of the token
before the token (typically a space token), instead of placing it just
before the next token, at the very beginning of the next line.

This patch handles this case to place the cursor rectangle correctly.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>