TextBox: fix longest_suffix()
When a 2-byte or more UTF-8 character was inserted into a right-aligned
TextBox, the program crashed throwing an exception with the message
"Invalid UTF-8".

This exception was risen from longest_suffix(). Indeed the
std::string::const_reverse_iterator was not suited as a parameter of the
utf8::iterator template.

Indeed, UTF-8 characters could not be iterated correctly.
For instance, with the "AB[C1C2]" string, [C1C2] being a 2-byte UTF-8
character, the reverse string should be "[C1C2]BA". However, iterating
with an utf8::interator<std::string::const_reverser_iterator> gave the
incorrect string "[C2C1]BA".

If [C1C2] is a well-formed UTF-8 character, [C2C1] can't be, hence the
thrown exception, and would not have been the expected UTF-8
character anyway.

So, use detail::utf8_const_iterator instead, that is to say an alias for
utf8::iterator<std::string::const_iterator>, and iterate from the end to
the beginning.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>